# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet & document management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type PrincipalRole = (typeof PrincipalRole)[keyof typeof PrincipalRole];

export const PrincipalRole = {
  admin: "admin",
  company: "company",
} as const;

export interface Principal {
  id: string;
  email: string;
  role: PrincipalRole;
  /** @nullable */
  companyId?: string | null;
  /** @nullable */
  companyName?: string | null;
}

export interface LoginResult {
  token: string;
  principal: Principal;
}

export interface MeResponse {
  principal: Principal;
}

export interface Company {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
  vehicleCount?: number;
}

export interface CreateCompanyRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  contactName?: string;
  phone?: string;
}

export interface CompanyListResponse {
  items: Company[];
  total: number;
  page: number;
  pageSize: number;
}

export type VehicleWorstStatus =
  (typeof VehicleWorstStatus)[keyof typeof VehicleWorstStatus];

export const VehicleWorstStatus = {
  valid: "valid",
  expiring: "expiring",
  expired: "expired",
  none: "none",
} as const;

export interface Vehicle {
  id: string;
  companyId: string;
  name: string;
  plateNumber: string;
  vinCode: string;
  year: number;
  /** @nullable */
  techPassportSeries?: string | null;
  /** @nullable */
  driverName?: string | null;
  createdAt: string;
  documentCount?: number;
  worstStatus?: VehicleWorstStatus;
  /** @nullable */
  nextExpiryAt?: string | null;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  valid: "valid",
  expiring: "expiring",
  expired: "expired",
} as const;

export interface Document {
  id: string;
  vehicleId: string;
  companyId: string;
  /** @nullable */
  vehicleName?: string | null;
  /** @nullable */
  vehiclePlateNumber?: string | null;
  name: string;
  number: string;
  startDate: string;
  endDate: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  status: DocumentStatus;
  daysRemaining: number;
  createdAt: string;
}

export type VehicleDetail = Vehicle & {
  documents: Document[];
};

export interface CreateVehicleRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  plateNumber: string;
  /** @minLength 1 */
  vinCode: string;
  /**
   * @minimum 1900
   * @maximum 2100
   */
  year: number;
  techPassportSeries?: string;
  driverName?: string;
}

export interface UpdateVehicleRequest {
  name?: string;
  plateNumber?: string;
  vinCode?: string;
  year?: number;
  techPassportSeries?: string;
  driverName?: string;
}

export interface VehicleListResponse {
  items: Vehicle[];
  total: number;
  page: number;
  pageSize: number;
}

export interface CreateDocumentRequest {
  vehicleId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  number: string;
  startDate: string;
  endDate: string;
  note?: string;
  fileUrl?: string;
  fileName?: string;
}

export interface UpdateDocumentRequest {
  name?: string;
  number?: string;
  startDate?: string;
  endDate?: string;
  note?: string;
  fileUrl?: string;
  fileName?: string;
}

export interface DocumentListResponse {
  items: Document[];
  total: number;
  page: number;
  pageSize: number;
}

export interface DashboardSummary {
  totalVehicles: number;
  totalDocuments: number;
  validCount: number;
  expiringCount: number;
  expiredCount: number;
  upcomingExpirations: Document[];
  recentlyExpired: Document[];
}

export type NotificationKind =
  (typeof NotificationKind)[keyof typeof NotificationKind];

export const NotificationKind = {
  expiring_soon: "expiring_soon",
  expired: "expired",
} as const;

export interface Notification {
  id: string;
  documentId: string;
  kind: NotificationKind;
  message: string;
  createdAt: string;
  /** @nullable */
  documentName?: string | null;
  /** @nullable */
  vehicleName?: string | null;
}

export interface NotificationListResponse {
  items: Notification[];
}

export interface RequestUploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResult {
  uploadURL: string;
  objectPath: string;
}

export type ListCompaniesParams = {
  search?: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};

export type ListVehiclesParams = {
  search?: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};

export type ListDocumentsParams = {
  search?: string;
  status?: ListDocumentsStatus;
  vehicleId?: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};

export type ListDocumentsStatus =
  (typeof ListDocumentsStatus)[keyof typeof ListDocumentsStatus];

export const ListDocumentsStatus = {
  valid: "valid",
  expiring: "expiring",
  expired: "expired",
} as const;