# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fleet & document management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated principal
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Principal info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyListResponse" }
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company with login credentials (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCompanyRequest" }
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
  /companies/{id}:
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for the current company
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [valid, expiring, expired]
        - in: query
          name: vehicleId
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list of documents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentListResponse" }
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document attached to a vehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDocumentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDocumentRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    LoginResult:
      type: object
      required: [token, principal]
      properties:
        token: { type: string }
        principal: { $ref: "#/components/schemas/Principal" }
    Principal:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role:
          type: string
          enum: [admin, company]
        companyId:
          type: [string, "null"]
          format: uuid
        companyName:
          type: [string, "null"]
    MeResponse:
      type: object
      required: [principal]
      properties:
        principal: { $ref: "#/components/schemas/Principal" }
    Company:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        contactName:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        createdAt: { type: string, format: date-time }
        vehicleCount:
          type: integer
    CreateCompanyRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        contactName: { type: string }
        phone: { type: string }
    CompanyListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Company" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    Document:
      type: object
      required:
        - id
        - vehicleId
        - companyId
        - name
        - number
        - startDate
        - endDate
        - status
        - daysRemaining
        - createdAt
      properties:
        id: { type: string, format: uuid }
        vehicleId: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        vehicleName:
          type: [string, "null"]
        vehiclePlateNumber:
          type: [string, "null"]
        name: { type: string }
        number: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        note:
          type: [string, "null"]
        fileUrl:
          type: [string, "null"]
        fileName:
          type: [string, "null"]
        status:
          type: string
          enum: [valid, expiring, expired]
        daysRemaining: { type: integer }
        createdAt: { type: string, format: date-time }
    CreateDocumentRequest:
      type: object
      required: [vehicleId, name, number, startDate, endDate]
      properties:
        vehicleId: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
        number: { type: string, minLength: 1 }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        note: { type: string }
        fileUrl: { type: string }
        fileName: { type: string }
    UpdateDocumentRequest:
      type: object
      properties:
        name: { type: string }
        number: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        note: { type: string }
        fileUrl: { type: string }
        fileName: { type: string }
    DocumentListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Document" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...