# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet & document management API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: Date;
  vehicleCount?: number;
}