# lib/api-zod/src/generated/types/companyListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet & document management API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from "./company";

export interface CompanyListResponse {
  items: Company[];
  total: number;
  page: number;
  pageSize: number;
}